Use long long instead of long
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@132455 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 5c054e23fc7c6e45478a83fbaa5cdf3711198e76

+++ Rec/Rich/RichAlignment/src/RichAlignmentEventSelection.cpp
@@ -140,7 +140,7 @@ StatusCode EventSelection::execute()
       if ( !unAmbiguousPhoton ) { continue; }
 
       // mirror combination number
-      const unsigned long int key = (100000*sphMirNum) + flatMirNum;
+      const unsigned long long key = (100000*sphMirNum) + flatMirNum;
 
       // count the photons in each of the mirror combinations
       ++m_combiMap[key];
